requirements.txt
discord.py==2.5.0
pyyaml==6.0.2
aiosqlite==0.21.0